refactor: service만 다시 분할
