+++ src/main/java/kcs/funding/fundingboost/domain/service/pay/MyPayService.java
@@ -1,0 +1,10 @@
+package kcs.funding.fundingboost.domain.service.pay;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MyPayService {
+
+}
